@@ -14,34 +14,34 @@
     "publish:lib": "npm run build:lib && npm publish ./dist"
   },
   "dependencies": {
-    "@angular/animations": "8.2.14",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/forms": "8.2.14",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
+    "@angular/animations": "9.0.5",
+    "@angular/common": "9.0.5",
+    "@angular/compiler": "9.0.5",
+    "@angular/core": "9.0.5",
+    "@angular/forms": "9.0.5",
+    "@angular/platform-browser": "9.0.5",
+    "@angular/platform-browser-dynamic": "9.0.5",
+    "@angular/router": "9.0.5",
     "core-js": "^2.4.1",
     "date-fns": "^1.30.1",
     "ngx-bootstrap": "^4.0.0",
     "rxjs": "^6.5.4",
     "ts-md5": "^1.2.4",
-    "tslib": "^1.9.3",
-    "zone.js": "~0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.25",
-    "@angular/cli": "8.3.25",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/language-service": "8.2.14",
+    "@angular-devkit/build-angular": "~0.900.5",
+    "@angular/cli": "9.0.5",
+    "@angular/compiler-cli": "9.0.5",
+    "@angular/language-service": "9.0.5",
     "@braune-digital/form-utilities": "^1.1.0",
     "@ng-select/ng-select": "^2.16.3",
     "@ngx-translate/core": "^11.0.1",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^5.0.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
@@ -49,7 +49,7 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^5.4.0",
+    "ng-packagr": "^9.0.0",
     "ng2-ckeditor": "^1.2.2",
     "ngx-quill": "^4.4.2",
     "ngx-textarea-autosize": "^2.0.0",
@@ -58,6 +58,6 @@
     "rimraf": "^2.6.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "^3.5.3"
+    "typescript": "^3.7.5"
   }
 }
